fix - - create interfaces for models, vendors and github repo - change testing framwork
