+++ backend-app/.mocharc.json
@@ -1,0 +1,5 @@
+{
+    "extension": ["ts"],
+    "spec": "tests/**/*.spec.ts",
+    "require": "tsconfig-paths/register"
+}
