--- backend-app/globalConfig.json
@@ -1,1 +1,0 @@
-{ "mongoUri": "mongodb://127.0.0.1:58466/" }
