--- backend-app/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    globalSetup: './tests/setup.js',
-    globalTeardown: './tests/teardown.js',
-    setupFilesAfterEnv: ['./tests/db_config.js'],
-    bail: true,
-};
