+++ backend-app/models/calendar/calendar_model.ts
@@ -1,24 +1,6 @@
 import mongoose, { Schema, Document } from 'mongoose';
 import metaData from '@constants/meta_data';
-
-export interface ICalendar extends Document {
-    Name: string;
-    Type: string;
-    isPublic: boolean;
-    isShareAble: boolean;
-    participants: mongoose.Schema.Types.ObjectId[];
-    events: mongoose.Schema.Types.ObjectId[];
-    description?: string;
-    accessCode?: string;
-    tags: string[];
-    allowedUsers: mongoose.Schema.Types.ObjectId[];
-    deniedUsers: mongoose.Schema.Types.ObjectId[];
-    deleted: boolean;
-    deletedBy?: string;
-    deletedAt?: Date;
-    createdBy?: string;
-    updatedBy?: string;
-}
+import { ICalendar } from '@interfaces/models/i_calendar';
 
 const calendarSchema: Schema = new Schema<ICalendar>(
     {
