+++ backend-app/models/calendar/event_model.ts
@@ -1,25 +1,6 @@
-import mongoose, { Schema, Document } from 'mongoose';
+import mongoose, { Schema } from 'mongoose';
 import metaData from '@constants/meta_data';
-
-export interface IEvent extends Document {
-    name: string;
-    description?: string;
-    location?: string;
-    startDate: Date;
-    endDate: Date;
-    startTime: string;
-    endTime: string;
-    color: string;
-    recurring?: boolean;
-    recurringType?: 'Daily' | 'Weekly' | 'Monthly' | 'Yearly';
-    recurringEndDate?: Date;
-    reminder?: Date;
-    deleted: boolean;
-    deletedBy?: string;
-    deletedAt?: Date;
-    createdBy?: string;
-    updatedBy?: string;
-}
+import { IEvent } from '@interfaces/models/i_event';
 
 const eventSchema: Schema = new Schema<IEvent>({
     name: {
