+++ backend-app/models/user/user_model.ts
@@ -1,38 +1,10 @@
-import mongoose, { Document, Model, Schema } from 'mongoose';
+import mongoose, { Model, Schema } from 'mongoose';
 import validator from 'validator';
 import bcrypt from 'bcrypt';
 import Actions from '@constants/actions';
 import metaData from '@constants/meta_data';
 import { randomBytes, createHash } from 'crypto';
-
-export interface IUser extends Document {
-    name: string;
-    email: string;
-    address?: string;
-    password?: string;
-    authorities: Actions[];
-    restrictions: Actions[];
-    roles: string[];
-    active: boolean;
-    activationKey?: string;
-    accessRestricted: boolean;
-    githubOauthAccessToken?: string;
-    resetKey?: string;
-    createdAt: Date;
-    updatedAt: Date;
-    deleted: boolean;
-    deletedBy?: string;
-    deletedAt?: Date;
-    createdBy?: string;
-    updatedBy?: string;
-    correctPassword(
-        typedPassword: string,
-        originalPassword: string
-    ): Promise<boolean>;
-    isAuthorizedTo(action: Actions): boolean;
-    isRestrictedFrom(action: Actions): boolean;
-    generateResetKey(): string;
-}
+import { IUser } from '@interfaces/models/i_user';
 
 const userSchema: Schema = new mongoose.Schema<IUser>(
     {
