+++ backend-app/package.json
@@ -8,32 +8,25 @@
         "dev": "ts-node-dev server.ts",
         "prod": "node server.ts NODE_ENV=PRODUCTION",
         "start": "ts-node-dev server.ts",
-        "test": "jest --detectOpenHandles",
+        "test": "ts-mocha",
         "prepare": "cd .. && husky install",
         "seed": "ts-node-dev ./seeds/m.js"
     },
     "author": "Bella Abdelouahab",
     "license": "ISC",
     "dependencies": {
-        "@inquirer/prompts": "3.1.1",
-        "@typegoose/typegoose": "11.5.0",
-        "@types/compression": "1.7.3",
-        "@types/express-routes-versioning": "1.0.1",
-        "@types/hpp": "0.2.2",
-        "@types/morgan": "1.9.5",
-        "@types/validator": "13.11.1",
-        "@typescript-eslint/parser": "6.5.0",
         "axios": "1.5.0",
         "bcrypt": "5.1.1",
         "compression": "1.7.4",
         "cookie-parser": "^1.4.5",
         "cors": "^2.8.5",
         "dotenv": "16.3.1",
-        "express": "^4.17.1",
+        "express": "4.18.2",
         "express-bearer-token": "2.4.0",
         "express-mongo-sanitize": "2.2.0",
         "express-rate-limit": "6.10.0",
         "express-routes-versioning": "1.0.1",
+        "express-swagger-generator": "1.1.17",
         "helmet": "4.6.0",
         "hpp": "0.2.3",
         "http-status-codes": "2.2.0",
@@ -52,20 +45,31 @@
         "yamljs": "^0.3.0"
     },
     "devDependencies": {
+        "@inquirer/prompts": "3.1.1",
+        "@typegoose/typegoose": "11.5.0",
         "@types/bcrypt": "5.0.0",
         "@types/body-parser": "^1.19.0",
-        "@types/cookie-parser": "^1.4.2",
-        "@types/cors": "^2.8.10",
-        "@types/express": "^4.17.13",
-        "@types/helmet": "^4.0.0",
+        "@types/chai": "4.3.6",
+        "@types/compression": "1.7.3",
+        "@types/cookie-parser": "1.4.4",
+        "@types/cors": "2.8.14",
+        "@types/expect": "24.3.0",
+        "@types/express": "4.17.17",
+        "@types/express-routes-versioning": "1.0.1",
+        "@types/hpp": "0.2.3",
+        "@types/mocha": "10.0.1",
+        "@types/morgan": "1.9.5",
         "@types/node": "^16.11.7",
+        "@types/validator": "13.11.1",
+        "@typescript-eslint/parser": "6.7.2",
+        "chai": "4.3.8",
         "eslint": "8.48.0",
         "eslint-config-prettier": "9.0.0",
         "eslint-plugin-prettier": "5.0.0",
         "husky": "7.0.4",
-        "jest": "27.5.1",
         "lint-staged": "14.0.1",
         "prettier": "3.0.3",
+        "ts-mocha": "10.0.0",
         "ts-node": "^10.4.0",
         "tsconfig-paths": "4.2.0",
         "typescript": "5.2.2"
