+++ backend-app/routes/index.ts
@@ -4,14 +4,14 @@ import adminRoutes from './users/admin_route';
 import superAdminRoutes from './users/super_admin_route';
 import authRoutes from './auth_routes';
 import githubRoutes from './github_routes';
-import authController from '../controllers/auth_controllers/auth_controller';
+import { protect } from '../controllers/auth_controllers/auth_controller';
 
 const router = Router();
 
 // public routes
 authRoutes(router);
 
-router.use(authController.protect);
+router.use(protect);
 
 // protected routes
 userRoutes(router);
