+++ backend-app/tests/init/roles.spec.ts
@@ -1,8 +1,15 @@
+// imprt expect from chai
+
+import '../../utils/register_paths';
+
+import expect from 'expect';
+
 describe('Role tests', () => {
-    test('should create roles', async () => {
+    it('should create roles', () => {
         // await createRoles();
         // const response = await request(app).post('/ap/auth/signup');
         // console.log(response.body);
+        expect(1).toBe(1);
     });
     it('should create default user', async () => {
         // await createDefaultUser();
