+++ backend-app/tests/init/x.spec.ts
@@ -1,0 +1,11 @@
+import { describe, it } from 'mocha';
+// @ts-ignore
+import logger from '@utils/logger';
+
+logger.info('Hello World!');
+
+describe('Your test suite', () => {
+    it('Your test case', () => {
+        // Your test assertion goes here
+    });
+});
